[sortinghat] Add dedicated workers/queues
For BAP 0.19.0/GrimoireLab 0.22.0, the support for
dedicated queues on SortingHat was added. Now on,
tenants can used dedicated queues to improve the
performance and security of identities jobs.

This commit adds that support, allowing the creation
of dedicated workers for the tenants that require
this configuration. With this change, there will be
two types of workers: normal workers, that will get
the jobs from the 'default' queue; dedicated workers,
one worker for each tenant that will only get jobs from
queue with the same tenant name.

Some changes were also required on the ansible config.
There's a new entry names 'sortinghat' under the instances
var. This entry will store all the specific configuration
related to each SortingHat instance. For example, OpenInfra
tokens will be now specific to each instance and not
shared across all the other instances.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>