[nginx] Improve certbot renew tasks
This code creates new certificates if they do not exist and the
crontab script checks daily if the certificates are still valid
and renew them if necessary. By default, if the expiration time
is less than 30 days, it will be renewed (it can be changed by
setting the `nginx_letsencrypt_cert_warning_days` parameter).

Signed-off-by: Quan Zhou <quan@bitergia.com>